Added version 1.1.453

+++ chromium/chrome/android/java/src/org/chromium/chrome/browser/ntp/RecentTabsManager.java
@@ -5,7 +5,9 @@
 package org.chromium.chrome.browser.ntp;
 
 import android.content.Context;
+import android.content.SharedPreferences;
 import android.graphics.Bitmap;
+import android.preference.PreferenceManager;
 
 import org.chromium.base.ObserverList;
 import org.chromium.base.ThreadUtils;
@@ -52,6 +54,8 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
     }
 
     private static final int RECENTLY_CLOSED_MAX_TAB_COUNT = 5;
+    private static final String PREF_SIGNIN_PROMO_DECLINED =
+            "recent_tabs_signin_promo_declined";
 
     private final Profile mProfile;
     private final Tab mTab;
@@ -219,6 +223,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      */
     public void openForeignSessionTab(ForeignSession session, ForeignSessionTab tab,
             int windowDisposition) {
+        if (mIsDestroyed) return;
         NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_FOREIGN_SESSION);
         mForeignSessionHelper.openForeignSessionTab(mTab, session, tab, windowDisposition);
     }
@@ -231,6 +236,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      *         be restored into the current tab or a new tab.
      */
     public void openRecentlyClosedTab(RecentlyClosedTab tab, int windowDisposition) {
+        if (mIsDestroyed) return;
         NewTabPageUma.recordAction(NewTabPageUma.ACTION_OPENED_RECENTLY_CLOSED_ENTRY);
         mRecentlyClosedBridge.openRecentlyClosedTab(mTab, tab, windowDisposition);
     }
@@ -239,6 +245,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Opens the history page.
      */
     public void openHistoryPage() {
+        if (mIsDestroyed) return;
         mTab.loadUrl(new LoadUrlParams(UrlConstants.HISTORY_URL));
         StartupMetrics.getInstance().recordOpenedHistory();
     }
@@ -283,6 +290,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the currently open tabs list is collapsed.
      */
     public void setCurrentlyOpenTabsCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setCurrentlyOpenTabsCollapsed(isCollapsed);
     }
 
@@ -326,6 +334,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the session is collapsed or expanded.
      */
     public void setForeignSessionCollapsed(ForeignSession session, boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setForeignSessionCollapsed(session, isCollapsed);
     }
 
@@ -346,6 +355,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the recently closed tabs list is collapsed.
      */
     public void setRecentlyClosedTabsCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setRecentlyClosedTabsCollapsed(isCollapsed);
     }
 
@@ -366,6 +376,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param session Session to be deleted.
      */
     public void deleteForeignSession(ForeignSession session) {
+        if (mIsDestroyed) return;
         mForeignSessionHelper.deleteForeignSession(session);
     }
 
@@ -373,6 +384,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * Clears the list of recently closed tabs.
      */
     public void clearRecentlyClosedTabs() {
+        if (mIsDestroyed) return;
         mRecentlyClosedBridge.clearRecentlyClosedTabs();
     }
 
@@ -386,7 +398,22 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
             return false;
         }
 
+        if (PreferenceManager.getDefaultSharedPreferences(mContext).getBoolean(
+                PREF_SIGNIN_PROMO_DECLINED, false)) {
+            return false;
+        }
+
         return !AndroidSyncSettings.isSyncEnabled(mContext) || mForeignSessions.isEmpty();
+    }
+
+    /**
+     * Save that user tapped "No" button on the signin promo.
+     */
+    public void setSigninPromoDeclined() {
+        SharedPreferences.Editor sharedPreferencesEditor =
+                PreferenceManager.getDefaultSharedPreferences(mContext).edit();
+        sharedPreferencesEditor.putBoolean(PREF_SIGNIN_PROMO_DECLINED, true);
+        sharedPreferencesEditor.apply();
     }
 
     /**
@@ -395,6 +422,7 @@ public class RecentTabsManager implements AndroidSyncSettingsObserver, SignInSta
      * @param isCollapsed Whether the sync promo is collapsed.
      */
     public void setSyncPromoCollapsed(boolean isCollapsed) {
+        if (mIsDestroyed) return;
         mNewTabPagePrefs.setSyncPromoCollapsed(isCollapsed);
     }
